+++ xdef/src/test/java/test/xdutils/TestGUIDebuger.java
@@ -1,4 +1,4 @@
-package test.xdef;
+package test.xdutils;
 
 import test.XDTester;
 import org.xdef.XDConstants;
@@ -21,9 +21,9 @@ import static test.XDTester._xdNS;
 /** Provides testing of XDef debug mode and editing in display mode.
  * @author Trojan
  */
-public final class TestDebugGUI extends XDTester {
+public final class TestGUIDebuger extends XDTester {
 
-	public TestDebugGUI() {super(); setChkSyntax(false);}
+	public TestGUIDebuger() {super(); setChkSyntax(false);}
 
 	@Override
 	public void test() {
@@ -50,8 +50,7 @@ public final class TestDebugGUI extends XDTester {
 		setProperty(XDConstants.XDPROPERTY_WARNINGS, //xdef_warnings
 			XDConstants.XDPROPERTYVALUE_WARNINGS_TRUE); //true
 //			XDConstants.XDPROPERTYVALUE_WARNINGS_FALSE); //false
-		try {
-			// XScript breakpoints
+		try { // XScript breakpoints
 			xdef =
 "<x:collection xmlns:x='" + _xdNS + "'>\n"+                 	//01
 "<x:def name = 'a' root = 'a'>\n"+								//02
@@ -109,8 +108,7 @@ public final class TestDebugGUI extends XDTester {
 			}
 			parse(xd, xml, reporter);
 		} catch (Exception ex) {fail(ex);}
-		try {
-			// XPos breakpoints
+		try { // XPos breakpoints
 			xdef =
 "<x:def xmlns:x='" + _xdNS + "' root='a'>\n"+
 "  <a a = \"required;\">\n"+
@@ -352,7 +350,7 @@ public final class TestDebugGUI extends XDTester {
 				System.out.println(reporter);
 			}
 		} catch (RuntimeException ex) {fail(ex);}
-
+		clearTempDir(); // clear temporary directory
 		resetTester();
 	}